Fixed error related to IoC container due to some changes made in previous PR
