+++ src/inversify.types.ts
@@ -1,0 +1,9 @@
+export const TYPES = {
+    IDifficultyRepo: Symbol.for("IDifficultyRepo"),
+    ICategoryRepo: Symbol.for("ICategoryRepo"),
+    IQuestionRepo: Symbol.for("IQuestionRepo"),
+    IAnswerRepo: Symbol.for("IAnswerRepo"),
+    IUserRepo: Symbol.for("IUserRepo"),
+    IQuizRepo: Symbol.for("IQuizRepo"),
+    ISubmissionRepo: Symbol.for("ISubmissionRepo"),
+};
